Use float for fit result histogram.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@4599 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ eklm/simulation/include/FPGAFitter.h
@@ -46,12 +46,12 @@ namespace Belle2 {
     /**
      * FPGA fitter.
      * @param[in]  amp     Digital amplitude.
-     * @param[in]  fit     Digital fit result histogram.
+     * @param[in]  fit     Fit result histogram.
      * @param[in]  nPoints Number of points in amplitude array.
      * @param[out] par     Fit parameters.
      * @return Fit status.
      */
-    enum FPGAFitStatus FPGAFit(int* amp, int* fit, int nPoints,
+    enum FPGAFitStatus FPGAFit(int* amp, float* fit, int nPoints,
                                struct FPGAFitParams* par);
 
   }
